// src/app/models/masters/tax-filters.ts
export class TaxFilters { 
    id: number= -1;
    name: string= "";
    abbreviation: string= "";
    idCountry: number = -1;
    TaxTypeId: number = -1;
    TaxBaseTypeId: number = -1;
    TaxBaseId: number = -1;
    idTaxeTypeApplication: number = -1;
    active: number = -1;
    TaxBaseName: string = "";
}